+++ src/main/java/com/cinchapi/runway/DatabaseInterface.java
@@ -18,12 +18,15 @@ package com.cinchapi.runway;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 import com.cinchapi.concourse.DuplicateEntryException;
 import com.cinchapi.concourse.lang.Criteria;
 import com.cinchapi.concourse.lang.paginate.Page;
 import com.cinchapi.concourse.lang.sort.Order;
+import com.cinchapi.runway.util.Paging;
+import com.google.common.collect.Sets;
 
 /**
  * A {@link DatabaseInterface} provides methods for interacting with a database
@@ -86,6 +89,33 @@ public interface DatabaseInterface {
     }
 
     /**
+     * Return the number of {@link Records} in the {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter}
+     * 
+     * @param clazz
+     * @param filter
+     * @return the number of {@link Records} in {@code clazz} that match the
+     *         {@code criteria}.
+     */
+    public default <T extends Record> int count(Class<T> clazz,
+            Criteria criteria, Predicate<T> filter) {
+        return find(clazz, criteria, filter).size();
+    }
+
+    /**
+     * Return the number of {@link Records} in the {@code clazz} that pass the
+     * {@code filter}.
+     * 
+     * @param clazz
+     * @param filter
+     * @return the number of {@link Records} in {@code clazz}.
+     */
+    public default <T extends Record> int count(Class<T> clazz,
+            Predicate<T> filter) {
+        return load(clazz, filter).size();
+    }
+
+    /**
      * Return the number of {@link Records} across the hierarchy of
      * {@code clazz}.
      * 
@@ -93,7 +123,7 @@ public interface DatabaseInterface {
      * @return the number of {@link Records} in {@code clazz}.
      */
     public default <T extends Record> int countAny(Class<T> clazz) {
-        return load(clazz).size();
+        return loadAny(clazz).size();
     }
 
     /**
@@ -101,12 +131,40 @@ public interface DatabaseInterface {
      * {@code clazz} that match the {@code criteria}.
      * 
      * @param clazz
+     * @param criteria
      * @return the number of {@link Records} in {@code clazz} that match the
      *         {@code criteria}.
      */
     public default <T extends Record> int countAny(Class<T> clazz,
             Criteria criteria) {
-        return find(clazz, criteria).size();
+        return findAny(clazz, criteria).size();
+    }
+
+    /**
+     * Return the number of {@link Records} across the hierarchy of
+     * {@code clazz} that match the {@code criteria}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param filter
+     * @return the number of {@link Records} in {@code clazz} that match the
+     *         {@code criteria}.
+     */
+    public default <T extends Record> int countAny(Class<T> clazz,
+            Criteria criteria, Predicate<T> filter) {
+        return findAny(clazz, criteria, filter).size();
+    }
+
+    /**
+     * Return the number of {@link Records} across the hierarchy of
+     * {@code clazz} that pass the {@code filter}
+     * 
+     * @param clazz
+     * @return the number of {@link Records} in {@code clazz}.
+     */
+    public default <T extends Record> int countAny(Class<T> clazz,
+            Predicate<T> filter) {
+        return loadAny(clazz, filter).size();
     }
 
     /**
@@ -144,7 +202,6 @@ public interface DatabaseInterface {
      * @param criteria
      * @param order
      * @return the matching records
-     * @deprecated Use {@link #find(Class, Criteria, Order)} instead
      */
     public <T extends Record> Set<T> find(Class<T> clazz, Criteria criteria,
             Order order);
@@ -162,6 +219,43 @@ public interface DatabaseInterface {
      */
     public <T extends Record> Set<T> find(Class<T> clazz, Criteria criteria,
             Order order, Page page);
+
+    /**
+     * Find and return all the records of type {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter}, sorted by the specified
+     * {@code order} and limited to the
+     * specified {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param order
+     * @param page
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> find(Class<T> clazz,
+            Criteria criteria, Order order, Page page, Predicate<T> filter) {
+        Set<T> unfiltered = find(clazz, criteria, order);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Find and return all the records of type {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter}, sorted by the specified
+     * {@code order}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param order
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> find(Class<T> clazz,
+            Criteria criteria, Order order, Predicate<T> filter) {
+        Set<T> unfiltered = find(clazz, criteria, order);
+        return Sets.filter(unfiltered, filter::test);
+    }
 
     /**
      * Find and return all the records of type {@code clazz} that match the
@@ -189,6 +283,56 @@ public interface DatabaseInterface {
     public default <T extends Record> Set<T> find(Class<T> clazz,
             Criteria criteria, Page page, Order order) {
         return find(clazz, criteria, order, page);
+    }
+
+    /**
+     * Find and return all the records of type {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter}, sorted by the specified
+     * {@code order} and limited to the specified {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param order
+     * @param page
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> find(Class<T> clazz,
+            Criteria criteria, Page page, Order order, Predicate<T> filter) {
+        return find(clazz, criteria, order, page, filter);
+    }
+
+    /**
+     * Find and return all the records of type {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter} limited to the specified
+     * {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param page
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> find(Class<T> clazz,
+            Criteria criteria, Page page, Predicate<T> filter) {
+        Set<T> unfiltered = find(clazz, criteria);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Find and return all the records of type {@code clazz} that match the
+     * {@code criteria} and pass the {@code filter}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> find(Class<T> clazz,
+            Criteria criteria, Predicate<T> filter) {
+        Set<T> unfiltered = find(clazz, criteria);
+        return Sets.filter(unfiltered, filter::test);
     }
 
     /**
@@ -263,6 +407,42 @@ public interface DatabaseInterface {
 
     /**
      * Execute the {@link #find(Class, Criteria)} query for {@code clazz}
+     * and all of its descendants that pass the {@code filter}, sorted by the
+     * specified {@code order} and limited to the specified {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param order
+     * @param page
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria, Order order, Page page, Predicate<T> filter) {
+        Set<T> unfiltered = findAny(clazz, criteria, order);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz}
+     * and all of its descendants that pass the {@code filter}, sorted by the
+     * specified {@code order}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param order
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria, Order order, Predicate<T> filter) {
+        Set<T> unfiltered = findAny(clazz, criteria, order);
+        return Sets.filter(unfiltered, filter::test);
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz}
      * and all of its descendants limited to the specified {@code page}.
      * 
      * @param clazz
@@ -287,6 +467,56 @@ public interface DatabaseInterface {
     public default <T extends Record> Set<T> findAny(Class<T> clazz,
             Criteria criteria, Page page, Order order) {
         return findAny(clazz, criteria, order, page);
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz}
+     * and all of its descendants that pass the {@code filter}, sorted by the
+     * specified {@code order} and limited to the specified {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param page
+     * @param order
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria, Page page, Order order, Predicate<T> filter) {
+        return findAny(clazz, criteria, order, page, filter);
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz}
+     * and all of its descendants that pass the {@code filter}, limited to the
+     * specified {@code page}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param page
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria, Page page, Predicate<T> filter) {
+        Set<T> unfiltered = findAny(clazz, criteria);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Execute the {@link #find(Class, Criteria)} query for {@code clazz} and
+     * all of its descendants and return those that pass the {@code filter}.
+     * 
+     * @param clazz
+     * @param criteria
+     * @param filter
+     * @return the matching records
+     */
+    public default <T extends Record> Set<T> findAny(Class<T> clazz,
+            Criteria criteria, Predicate<T> filter) {
+        Set<T> unfiltered = findAny(clazz, criteria);
+        return Sets.filter(unfiltered, filter::test);
     }
 
     /**
@@ -422,6 +652,54 @@ public interface DatabaseInterface {
 
     /**
      * Load all the Records that are contained within the specified
+     * {@code clazz} that pass the {@code filter}, sorted using the specified
+     * {@code order} and limited to the specified {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param order
+     * @param page
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> load(Class<T> clazz, Order order,
+            Page page, Predicate<T> filter) {
+        Set<T> unfiltered = load(clazz, order);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} that pass the {@code filter}, sorted using the specified
+     * {@code order}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param order
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> load(Class<T> clazz, Order order,
+            Predicate<T> filter) {
+        Set<T> unfiltered = load(clazz, order);
+        return Sets.filter(unfiltered, filter::test);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
      * {@code clazz} and limited to the specified {@code page}.
      * 
      * <p>
@@ -457,6 +735,73 @@ public interface DatabaseInterface {
     public default <T extends Record> Set<T> load(Class<T> clazz, Page page,
             Order order) {
         return load(clazz, order, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} that pass the {@code filter}, sorted using the specified
+     * {@code order} and limited to the specified {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param page
+     * @param order
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> load(Class<T> clazz, Page page,
+            Order order, Predicate<T> filter) {
+        return load(clazz, order, page, filter);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} that pass the {@code filter}, limited to the specified
+     * {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param page
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> load(Class<T> clazz, Page page,
+            Predicate<T> filter) {
+        Set<T> unfiltered = load(clazz);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} and pass the {@code filter}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> load(Class<T> clazz,
+            Predicate<T> filter) {
+        Set<T> unfiltered = load(clazz);
+        return Sets.filter(unfiltered, filter::test);
     }
 
     /**
@@ -560,6 +905,55 @@ public interface DatabaseInterface {
 
     /**
      * Load all the Records that are contained within the specified
+     * {@code clazz} or any of its descendants that pass the {@code filter},
+     * sorted using the specified {@code order} and limited to the specified
+     * {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param order
+     * @param page
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> loadAny(Class<T> clazz,
+            Order order, Page page, Predicate<T> filter) {
+        Set<T> unfiltered = loadAny(clazz, order);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} or any of its descendants that pass the {@code filter},
+     * sorted using the specified {@code order}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param order
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> loadAny(Class<T> clazz,
+            Order order, Predicate<T> filter) {
+        Set<T> unfiltered = loadAny(clazz, order);
+        return Sets.filter(unfiltered, filter::test);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
      * {@code clazz} and limited to the specified {@code page}.
      * 
      * <p>
@@ -595,6 +989,75 @@ public interface DatabaseInterface {
     public default <T extends Record> Set<T> loadAny(Class<T> clazz, Page page,
             Order order) {
         return loadAny(clazz, order, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} or any of its descendants that pass the {@code filter},
+     * sorted using the specified {@code order} and limited to the specified
+     * {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param page
+     * @param order
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> loadAny(Class<T> clazz, Page page,
+            Order order, Predicate<T> filter) {
+        return loadAny(clazz, order, page, filter);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} that pass the {@code filter}, limited to the specified
+     * {@code page}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param page
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> loadAny(Class<T> clazz, Page page,
+            Predicate<T> filter) {
+        Set<T> unfiltered = loadAny(clazz);
+        Set<T> filtered = Sets.filter(unfiltered, filter::test);
+        return Paging.paginate(filtered, page);
+    }
+
+    /**
+     * Load all the Records that are contained within the specified
+     * {@code clazz} or any of its descendants and pass the {@code filter}.
+     * 
+     * <p>
+     * Multiple calls to this method with the same parameters will return
+     * <strong>different</strong> instances (e.g. the instances are not cached).
+     * This is done deliberately so different threads/clients can make changes
+     * to a Record in isolation.
+     * </p>
+     * 
+     * @param clazz
+     * @param filter
+     * @return a {@link Set set} of {@link Record} objects
+     */
+    public default <T extends Record> Set<T> loadAny(Class<T> clazz,
+            Predicate<T> filter) {
+        Set<T> unfiltered = loadAny(clazz);
+        return Sets.filter(unfiltered, filter::test);
     }
 
     /**