+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+#### Version 1.5.0 (November 17, 2019)
+* Fixed a bug that caused the `countAny` methods to return the wrong data.
+* Added methods to the `Runway` driver that support filtering data. Unlike a `Criteria` or `Condition` a `filter` is a `Predicate` that receives the loaded `Record` as input and executes business logic to determine whether the `Record` should be included in the result set. For example, filtering can be used to seamlessly enforce permissions in a `Runway` method call by passing in a predicate that checks whether the caller has access to the `Record`.
+* Remove support for record caching. The `Runway#builder#cache` method has been deprecated. Providing a record cache to Runway no longer has any effect. 
+* Added an `onLoad` hook to the `Record` class that can be used to provide a routine that is executed whenever an existing Record is loaded from the database.
+* Fixed a bug that caused the linked objects included in the `map` or `json` functions to not respect the provided `SerializationOptions`.
+
 #### Version 1.4.1 (October 2, 2019)
 * Fixed a regression bug where the `Runway#findAnyUnique` failed because an attempt was made to instantiate an object of the provided class instead of the record's stored class.